DataMaintaner实现读取范围内RuntimeStatusData的方法，更新获取TestInstance的接口返回long类型

+++ source/src/Modules/DataMaintainer/RuntimeDatabaseProxy.cs
@@ -10,7 +10,7 @@ namespace Testflow.DataMaintainer
         {
         }
 
-        public override int GetTestInstanceCount(string filterString)
+        public override long GetTestInstanceCount(string filterString)
         {
             throw new System.InvalidOperationException();
         }
